refactor: event details component
